@dongjoon-hyun 2020-04-11
Hi, All.
Unfortunately, the original `SQLQueryTestSuite` commit broke this `ThriftServerQueryTestSuite` in Maven environment in `master` and `branch-3.0`.
- https://amplab.cs.berkeley.edu/jenkins/view/Spark%20QA%20Test%20(Dashboard)/job/spark-branch-3.0-test-maven-hadoop-2.7-hive-2.3/211/

Please see the follow-up fix.
- https://github.com/apache/spark/pull/28186